Home screen animation And after 4 seconds switch to the Log In screen
