+++ app/src/main/java/com/example/workingdog/HomeActivity.java
@@ -2,13 +2,49 @@ package com.example.workingdog;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
+import android.view.WindowManager;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
+import android.widget.ImageView;
+import android.widget.TextView;
+
+import com.example.workingdog.ui.login.LoginActivity;
 
 public class HomeActivity extends AppCompatActivity {
+
+    private static final int SPLASH_TIME_OUT = 4000;
+
+    //variables
+    Animation topAnim, bottomAnim;
+    ImageView image;
+    TextView welcomeText ;
+
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
+        //getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
+        setContentView(R.layout.activity_home);
+        //Animation
+        topAnim = AnimationUtils.loadAnimation(this,R.anim.top_animation);
+        bottomAnim = AnimationUtils.loadAnimation(this,R.anim.bottom_animation);
+        //hooks
+        image = findViewById(R.id.imageView3);
+        welcomeText = findViewById(R.id.welcomeText);
+
+        image.setAnimation(bottomAnim);
+        welcomeText.setAnimation(topAnim);
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent homeIntent = new Intent(HomeActivity.this, LoginActivity.class);
+                startActivity(homeIntent);
+                finish();
+            }
+        },SPLASH_TIME_OUT);
     }
 }
