+++ app/src/main/java/com/example/workingdog/HomeActivity2.java
@@ -1,0 +1,14 @@
+package com.example.workingdog;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class HomeActivity2 extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_home2);
+    }
+}
