--- app/src/main/java/com/example/workingdog/MainActivityHome.java
@@ -1,27 +1,0 @@
-package com.example.workingdog;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-
-public class MainActivityHome extends AppCompatActivity {
-    private static int SPLASH_TIME_OUT = 3000 ;
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent homeIntent = new Intent(MainActivityHome.this, HomeActivity.class);
-                startActivity(homeIntent);
-                finish();
-            }
-        },SPLASH_TIME_OUT);
-    }
-
-}
