+++ app/src/main/res/values/preloaded_fonts.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <array name="preloaded_fonts" translatable="false">
+        <item>@font/aclonica</item>
+        <item>@font/antic</item>
+        <item>@font/bangers</item>
         <item>@font/roboto</item>
     </array>
 </resources>
